# .sisyphus/openapi.json
{"openapi":"3.0.1","info":{"title":"OpenAPI definition","version":"v0"},"servers":[{"url":"http://127.0.0.1:8080","description":"Generated server url"}],"paths":{"/api/v1/posts/{id}":{"put":{"tags":["post-controller"],"operationId":"update","parameters":[{"name":"id","in":"path","required":true,"schema":{"type":"integer","format":"int64"}}],"requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/UpdateRequest"}}},"required":true},"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/ApiResponseObject"}}}}}},"delete":{"tags":["post-controller"],"operationId":"archive","parameters":[{"name":"id","in":"path","required":true,"schema":{"type":"integer","format":"int64"}}],"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/ApiResponseObject"}}}}}}},"/api/v1/posts":{"get":{"tags":["post-controller"],"operationId":"list","parameters":[{"name":"limit","in":"query","required":false,"schema":{"type":"integer","format":"int32","default":20}},{"name":"offset","in":"query","required":false,"schema":{"type":"integer","format":"int32","default":0}}],"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/ApiResponseListPostSummary"}}}}}},"post":{"tags":["post-controller"],"operationId":"create","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/CreateRequest"}}},"required":true},"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/ApiResponseCreateResponse"}}}}}}},"/api/v1/auth/register":{"post":{"tags":["auth-controller"],"operationId":"register","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/RegisterRequest"}}},"required":true},"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/ApiResponseRegisterResponse"}}}}}}},"/api/v1/auth/refresh":{"post":{"tags":["auth-controller"],"operationId":"refresh","parameters":[{"name":"refresh_token","in":"cookie","required":false,"schema":{"type":"string"}}],"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/ApiResponseLoginResponse"}}}}}}},"/api/v1/auth/logout":{"post":{"tags":["auth-controller"],"operationId":"logout","parameters":[{"name":"refresh_token","in":"cookie","required":false,"schema":{"type":"string"}}],"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/ApiResponseObject"}}}}}}},"/api/v1/auth/login":{"post":{"tags":["auth-controller"],"operationId":"login","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/LoginRequest"}}},"required":true},"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/ApiResponseLoginResponse"}}}}}}},"/api/v1/tags":{"get":{"tags":["tag-controller"],"operationId":"listTags","responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/ApiResponseListTagCount"}}}}}}},"/api/v1/tags/{tag}":{"get":{"tags":["tag-controller"],"operationId":"listByTag","parameters":[{"name":"tag","in":"path","required":true,"schema":{"type":"string"}},{"name":"limit","in":"query","required":false,"schema":{"type":"integer","format":"int32","default":20}},{"name":"offset","in":"query","required":false,"schema":{"type":"integer","format":"int32","default":0}}],"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/ApiResponseListPostSummaryView"}}}}}}},"/api/v1/posts/{slug}":{"get":{"tags":["post-controller"],"operationId":"get","parameters":[{"name":"slug","in":"path","required":true,"schema":{"type":"string"}}],"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/ApiResponsePostDetail"}}}}}}},"/api/v1/health":{"get":{"tags":["health-controller"],"operationId":"health","responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/ApiResponseMapStringObject"}}}}}}},"/api/v1/categories":{"get":{"tags":["tag-controller"],"operationId":"listCategories","responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/ApiResponseListCategoryCount"}}}}}}},"/api/v1/authors/{id}":{"get":{"tags":["author-controller"],"operationId":"getAuthor","parameters":[{"name":"id","in":"path","required":true,"schema":{"type":"integer","format":"int64"}}],"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/ApiResponseAuthorView"}}}}}}}},"components":{"schemas":{"UpdateRequest":{"type":"object","properties":{"title":{"type":"string"},"summary":{"type":"string"},"contentMarkdown":{"type":"string"},"tagsCsv":{"type":"string"},"category":{"type":"string"},"status":{"type":"string"},"format":{"type":"string"}}},"ApiError":{"type":"object","properties":{"code":{"type":"string"},"message":{"type":"string"}}},"ApiResponseObject":{"type":"object","properties":{"success":{"type":"boolean"},"data":{"type":"object"},"error":{"$ref":"#/components/schemas/ApiError"}}},"CreateRequest":{"type":"object","properties":{"title":{"type":"string"},"summary":{"type":"string"},"contentMarkdown":{"type":"string"},"tagsCsv":{"type":"string"},"category":{"type":"string"},"status":{"type":"string"},"format":{"type":"string"}}},"ApiResponseCreateResponse":{"type":"object","properties":{"success":{"type":"boolean"},"data":{"$ref":"#/components/schemas/CreateResponse"},"error":{"$ref":"#/components/schemas/ApiError"}}},"CreateResponse":{"type":"object","properties":{"postId":{"type":"integer","format":"int64"},"slug":{"type":"string"}}},"RegisterRequest":{"type":"object","properties":{"email":{"type":"string"},"password":{"type":"string"},"inviteCode":{"type":"string"},"emailCode":{"type":"string"}}},"ApiResponseRegisterResponse":{"type":"object","properties":{"success":{"type":"boolean"},"data":{"$ref":"#/components/schemas/RegisterResponse"},"error":{"$ref":"#/components/schemas/ApiError"}}},"RegisterResponse":{"type":"object","properties":{"userId":{"type":"integer","format":"int64"}}},"ApiResponseLoginResponse":{"type":"object","properties":{"success":{"type":"boolean"},"data":{"$ref":"#/components/schemas/LoginResponse"},"error":{"$ref":"#/components/schemas/ApiError"}}},"LoginResponse":{"type":"object","properties":{"accessToken":{"type":"string"}}},"LoginRequest":{"type":"object","properties":{"email":{"type":"string"},"password":{"type":"string"}}},"ApiResponseListTagCount":{"type":"object","properties":{"success":{"type":"boolean"},"data":{"type":"array","items":{"$ref":"#/components/schemas/TagCount"}},"error":{"$ref":"#/components/schemas/ApiError"}}},"TagCount":{"type":"object","properties":{"tag":{"type":"string"},"count":{"type":"integer","format":"int32"}}},"ApiResponseListPostSummaryView":{"type":"object","properties":{"success":{"type":"boolean"},"data":{"type":"array","items":{"$ref":"#/components/schemas/PostSummaryView"}},"error":{"$ref":"#/components/schemas/ApiError"}}},"PostSummaryView":{"type":"object","properties":{"id":{"type":"integer","format":"int64"},"authorId":{"type":"integer","format":"int64"},"title":{"type":"string"},"slug":{"type":"string"},"summary":{"type":"string"},"tags":{"type":"array","items":{"type":"string"}},"category":{"type":"string"},"createdAt":{"type":"string","format":"date-time"}}},"ApiResponseListPostSummary":{"type":"object","properties":{"success":{"type":"boolean"},"data":{"type":"array","items":{"$ref":"#/components/schemas/PostSummary"}},"error":{"$ref":"#/components/schemas/ApiError"}}},"PostSummary":{"type":"object","properties":{"id":{"type":"integer","format":"int64"},"authorId":{"type":"integer","format":"int64"},"title":{"type":"string"},"slug":{"type":"string"},"summary":{"type":"string"},"tagsCsv":{"type":"string"},"category":{"type":"string"},"createdAt":{"type":"string","format":"date-time"}}},"ApiResponsePostDetail":{"type":"object","properties":{"success":{"type":"boolean"},"data":{"$ref":"#/components/schemas/PostDetail"},"error":{"$ref":"#/components/schemas/ApiError"}}},"PostDetail":{"type":"object","properties":{"id":{"type":"integer","format":"int64"},"authorId":{"type":"integer","format":"int64"},"title":{"type":"string"},"slug":{"type":"string"},"summary":{"type":"string"},"contentMarkdown":{"type":"string"},"contentHtml":{"type":"string"},"format":{"type":"string"},"tagsCsv":{"type":"string"},"category":{"type":"string"},"createdAt":{"type":"string","format":"date-time"},"updatedAt":{"type":"string","format":"date-time"}}},"ApiResponseMapStringObject":{"type":"object","properties":{"success":{"type":"boolean"},"data":{"type":"object","additionalProperties":{"type":"object"}},"error":{"$ref":"#/components/schemas/ApiError"}}},"ApiResponseListCategoryCount":{"type":"object","properties":{"success":{"type":"boolean"},"data":{"type":"array","items":{"$ref":"#/components/schemas/CategoryCount"}},"error":{"$ref":"#/components/schemas/ApiError"}}},"CategoryCount":{"type":"object","properties":{"category":{"type":"string"},"count":{"type":"integer","format":"int32"}}},"ApiResponseAuthorView":{"type":"object","properties":{"success":{"type":"boolean"},"data":{"$ref":"#/components/schemas/AuthorView"},"error":{"$ref":"#/components/schemas/ApiError"}}},"AuthorView":{"type":"object","properties":{"id":{"type":"integer","format":"int64"},"nickname":{"type":"string"},"avatarUrl":{"type":"string"},"bio":{"type":"string"},"createdAt":{"type":"string","format":"date-time"}}}}}}